separated questions and title into two separate divs, made questions OL
I wanted my title and questions to be in two separate divs, for the purposes of assigning a clickListener to the questions Div later on. Also, I added these divs to the HTML and grabbed them as Vars in the script file. This minimizes the amount of elements being created inside the actual quiz function, which makes them able to be called upon in later functions if necessary.